requirements.txt
django-cors-headers
Pillow
django
djangorestframework
python-decouple